Move common.utils.accounts to users.utils
